+++ workflows_definition/src/openapi/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.9.0"
-    gen_version: str = "2.436.1"
-    user_agent: str = "speakeasy-sdk/python 1.9.0 2.436.1 1.0.0 openapi"
+    sdk_version: str = "1.9.1"
+    gen_version: str = "2.438.15"
+    user_agent: str = "speakeasy-sdk/python 1.9.1 2.438.15 1.0.0 openapi"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None